manage-api/requirements.txt
uvicorn[standard]==0.13.1
gunicorn==20.0.4
fastapi==0.62.0
pyyaml==5.3.1
requests==2.25.0
loguru==0.5.3